Fancied up the flair (2% chance of metallic flair), added letter flair to vending machine.

+++ Assets/Scripts/BurgerScoring/Coin.cs
@@ -2,5 +2,5 @@
 using System.Collections;
 
 public class Coin : MonoBehaviour {
-    public int coinValue = 1;
+    public float coinValue = 1;
 }
